pyproject.toml
[tool.poetry]
name = "starkware"
version = "0.1.0"
description = ""
authors = ["Tomasz Mierzwa <tomek@tfi.team>, Karol Chojnowski <karol@tfi.team>"]

[tool.poetry.dependencies]
python = "^3.9"
Flask = "^2.0.2"
web3 = "^5.24.0"
cairo-lang = "^0.4.2"
python-dotenv = "^0.19.1"


[tool.poetry.dev-dependencies]
pre-commit = "*"
autoflake = "^1.3.1"
flake8 = "^4.0.1"
black = "==19.3b0"

[tool.isort]
multi_line_output = 3
include_trailing_comma = true
force_grid_wrap = 0
line_length = 88

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
